Prototype difficulty calculation

+++ src/config/mod.rs
@@ -7,11 +7,11 @@ pub const SYMBOL: &str = "ZIK";
 pub const TOTAL_SUPPLY: u64 = 10000000000_000000000u64; // 10 Billion ZIK
 
 // Delta means: block size + state size changes
-pub const MAX_DELTA_SIZE: u32 = 1024 * 1024 * 1024; // Bytes
+pub const MAX_DELTA_SIZE: usize = 1024 * 1024 * 1024; // Bytes
 
 // Every n blocks, recalculate difficulty
-pub const BLOCK_TIME: u32 = 60; // Seconds
-pub const DIFFICULTY_CALC_INTERVAL: u32 = 128; // Blocks
+pub const BLOCK_TIME: usize = 60; // Seconds
+pub const DIFFICULTY_CALC_INTERVAL: usize = 128; // Blocks
 
 // 0 63 -> BAZUKA BASE KEY
 // 64 2111 -> hash(blk#0)
